restructured tests folder to reflect new src structure
