+++ src/documents/collections/collection.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { normalizeSort } from './utils';
+import { normalizeSort } from '../utils';
 import { FindCursor } from '@/src/documents/cursor';
 import {
   BulkWriteError, DataAPIDetailedErrorDescriptor,
