+++ tests/integration/administration/lifecycle.test.ts
