+++ tests/integration/documents/collections/count-documents.test.ts
