+++ tests/integration/documents/collections/delete-all.test.ts
