+++ tests/integration/documents/collections/delete-many.test.ts
