+++ tests/integration/documents/collections/delete-one.test.ts
