+++ tests/integration/documents/collections/estimated-document-count.test.ts
