+++ tests/integration/documents/collections/find-one-and-delete.test.ts
