+++ tests/integration/documents/collections/insert-many.test.ts
