+++ tests/integration/documents/collections/insert-one.test.ts
