+++ tests/integration/documents/collections/misc.test.ts
