+++ tests/integration/documents/collections/options.test.ts
