+++ tests/integration/documents/cursor.test.ts
