+++ tests/integration/documents/vectorize/vec-test-groups.ts
