+++ tests/integration/documents/vectorize/vec-whitelist-imp.ts
