+++ tests/integration/documents/vectorize/vectorize.test.ts
