+++ tests/integration/lib/api/fetch/fetch-native.test.ts
