+++ tests/unit/documents/collections/collection.test.ts
