+++ tests/unit/documents/cursor.test.ts
