+++ tests/unit/documents/embedding-providers.test.ts
