+++ tests/unit/documents/errors.test.ts
